Malacarne Example

+++ MixedExpression.cpp
@@ -5,8 +5,8 @@
 
 int main (){
 	
-	float total {};
-	float num1{}, num2 {}, num3 {};
+	int total {};
+	int num1{}, num2 {}, num3 {};
 	const int count {3};
 	
 	std::cout << "Enter 3 integers separated by spaces: " << std::endl;
@@ -21,7 +21,6 @@ int main (){
 	std::cout << "The 3 numbers were: " << num1 << ";" << num2 << ";" << num3 << std::endl;
 	std::cout << "The sum of the numbers is: " << total << std::endl;
 	std::cout << "The average of the numbers is: " << average << std::endl;
-	
 	
 	return 0;
 }
